+++ bench/run_gather.sh
@@ -87,7 +87,7 @@ do
     
     for DEFAULT_ALGO in "simple" "linear_sync" "binomial" "basic_linear"
     do        
-        LIBSWING_GATHER_ALGO_FAMILY="DEFAULT" ${MPIRUN} ${EXTRA_MCAS} -mca coll ^tbl -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} -mca coll_select_gather_algorithm ${DEFAULT_ALGO} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${n} ${iterations}
+        LIBSWING_GATHER_ALGO_FAMILY="DEFAULT" ${MPIRUN} ${EXTRA_MCAS} -mca coll ^tbi -mca coll_tuned_prealloc_size ${coll_tuned_prealloc_size} -mca coll_select_gather_algorithm ${DEFAULT_ALGO} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${n} ${iterations}
         ALGO_FNAME=default-$(echo ${DEFAULT_ALGO} | tr '_' '-')
         mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
         if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
@@ -102,34 +102,36 @@ do
     export LIBSWING_PREALLOC_SIZE=${PREALLOC_SIZE} 
     for PORTS in ${PORTS_LIST//,/ }
     do
-        export LIBSWING_NUM_PORTS=${PORTS}
-        # Run swing
-        export LIBSWING_GATHER_ALGO_FAMILY="SWING" 
-        export LIBSWING_GATHER_ALGO_LAYER="UTOFU" 
-        export LIBSWING_GATHER_ALGO="BINOMIAL_TREE_CONT_PERMUTE"    
-        for SEGMENT_SIZE in 0 #4096 65536 1048576
-        do                
-            if [ $SEGMENT_SIZE -lt $msg_size ]; then
-                LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${n} ${iterations}                    
-                ALGO_FNAME=${LIBSWING_GATHER_ALGO_FAMILY}-${LIBSWING_GATHER_ALGO}-${LIBSWING_GATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
-                mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
-                if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
-            fi
-        done
-    
-        # Run recdoub
-        export LIBSWING_GATHER_ALGO_FAMILY="RECDOUB" 
-        export LIBSWING_GATHER_ALGO_LAYER="UTOFU" 
-        export LIBSWING_GATHER_ALGO="BINOMIAL_TREE_CONT_PERMUTE"    
-        for SEGMENT_SIZE in 0 #4096 65536 1048576
-        do                
-            if [ $SEGMENT_SIZE -lt $msg_size ]; then
-                LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${n} ${iterations}                    
-                ALGO_FNAME=${LIBSWING_GATHER_ALGO_FAMILY}-${LIBSWING_GATHER_ALGO}-${LIBSWING_GATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
-                mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
-                if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
-            fi
-        done
+        if [ $n -ge $PORTS ]; then
+            export LIBSWING_NUM_PORTS=${PORTS}
+            # Run swing
+            export LIBSWING_GATHER_ALGO_FAMILY="SWING" 
+            export LIBSWING_GATHER_ALGO_LAYER="UTOFU" 
+            export LIBSWING_GATHER_ALGO="BINOMIAL_TREE_CONT_PERMUTE"    
+            for SEGMENT_SIZE in 0 #4096 65536 1048576
+            do                
+                if [ $SEGMENT_SIZE -lt $msg_size ]; then
+                    LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${n} ${iterations}                    
+                    ALGO_FNAME=${LIBSWING_GATHER_ALGO_FAMILY}-${LIBSWING_GATHER_ALGO}-${LIBSWING_GATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
+                    mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
+                    if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
+                fi
+            done
+        
+            # Run recdoub
+            export LIBSWING_GATHER_ALGO_FAMILY="RECDOUB" 
+            export LIBSWING_GATHER_ALGO_LAYER="UTOFU" 
+            export LIBSWING_GATHER_ALGO="BINOMIAL_TREE_CONT_PERMUTE"    
+            for SEGMENT_SIZE in 0 #4096 65536 1048576
+            do                
+                if [ $SEGMENT_SIZE -lt $msg_size ]; then
+                    LIBSWING_SEGMENT_SIZE=${SEGMENT_SIZE} ${MPIRUN} ${MPIRUN_MAP_BY_NODE_FLAG} ${MPIEXEC_OUT} -n ${p} ${MPIRUN_ADDITIONAL_FLAGS} ./bench ${COLLECTIVE} ${DATATYPE} ${n} ${iterations}                    
+                    ALGO_FNAME=${LIBSWING_GATHER_ALGO_FAMILY}-${LIBSWING_GATHER_ALGO}-${LIBSWING_GATHER_ALGO_LAYER}-${SEGMENT_SIZE}-${PORTS}
+                    mv ${OUT_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.csv; rm -f ${OUT_PREFIX}* 
+                    if [ -f ${ERR_PREFIX}*.0 ]; then mv ${ERR_PREFIX}*.0 ${OUTPUT_DIR}/${EXP_ID}/${n}_${ALGO_FNAME}_${DATATYPE_lc}.err; rm -f ${ERR_PREFIX}*; fi
+                fi
+            done
+        fi
     done
     echo " ${GREEN}[Done]${NC}"
 done
